Delete version info from footer
In the refactoring workshop we use a lockdown test that saves the (html)
output of an entire request. We use timecop to work around any changing
times and dates, but the version information in the footer contains a
timestamp that isn't affected by this because that is loaded before
timecop is triggered.

To avoid ugly workarounds, I'm simply removing it in this fork.